Added main function
Currently it's just a husk that can convert the solar rotations to different units on the Antikythera dials. If you have developed code for the project, please modify this file accordingly to implement your code. The solar rotations = 1 year = 365 days. If we are working with 365 1/4, please let me know and I can convert all the numbers for that system.